+++ src/parser/part/prelude.rs
@@ -4,8 +4,8 @@ pub(crate) use crate::ast::{
     Allocate, Bind, Connect, ConnectBody, ConnectionEnd, ConnectionUsageMember, DefinitionPrefix,
     ExhibitState, Expression, InOut, InterfaceUsage, InterfaceUsageBodyElement, Membership, Node,
     PartDef, PartDefBody, PartDefBodyElement, PartUsage, PartUsageBody, PartUsageBodyElement,
-    Perform, PerformBody, PerformBodyElement, PerformInOutBinding, RefBody, RefBodyElement,
-    RefDecl, VariantTypedUsage, VariantUsage,
+    Perform, PerformBody, PerformBodyElement, PerformInOutBinding, RefBody, RefDecl,
+    VariantTypedUsage, VariantUsage,
 };
 pub(crate) use crate::parser::action::{action_def, action_usage};
 pub(crate) use crate::parser::allocation::{allocation_def, allocation_usage};